+++ staker/events.go
@@ -2,19 +2,10 @@ package staker
 
 import (
 	cl "github.com/babylonlabs-io/btc-staker/babylonclient"
-	"github.com/btcsuite/btcd/btcec/v2"
-	"github.com/btcsuite/btcd/btcec/v2/schnorr"
-	"github.com/btcsuite/btcd/btcutil"
 	"github.com/btcsuite/btcd/chaincfg/chainhash"
 	"github.com/btcsuite/btcd/wire"
-	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/sirupsen/logrus"
 )
-
-type responseExpectedChan struct {
-	errChan     chan error
-	successChan chan *chainhash.Hash
-}
 
 type StakingEvent interface {
 	// Each staking event is identified by initial staking transaction hash
@@ -22,131 +13,13 @@ type StakingEvent interface {
 	EventDesc() string
 }
 
-var _ StakingEvent = (*stakingRequestedEvent)(nil)
 var _ StakingEvent = (*stakingTxBtcConfirmedEvent)(nil)
 var _ StakingEvent = (*delegationSubmittedToBabylonEvent)(nil)
 var _ StakingEvent = (*delegationActiveOnBabylonEvent)(nil)
 var _ StakingEvent = (*unbondingTxSignaturesConfirmedOnBabylonEvent)(nil)
 var _ StakingEvent = (*unbondingTxConfirmedOnBtcEvent)(nil)
 var _ StakingEvent = (*spendStakeTxConfirmedOnBtcEvent)(nil)
-var _ StakingEvent = (*sendStakingTxToBTCRequestedEvent)(nil)
 var _ StakingEvent = (*criticalErrorEvent)(nil)
-
-type stakingRequestedEvent struct {
-	stakerAddress           btcutil.Address
-	stakingTxHash           chainhash.Hash
-	stakingTx               *wire.MsgTx
-	stakingOutputIdx        uint32
-	stakingOutputPkScript   []byte
-	stakingTime             uint16
-	stakingValue            btcutil.Amount
-	fpBtcPks                []*btcec.PublicKey
-	requiredDepthOnBtcChain uint32
-	pop                     *cl.BabylonPop
-	watchTxData             *watchTxData
-	usePreApprovalFlow      bool
-	errChan                 chan error
-	successChan             chan *chainhash.Hash
-}
-
-func (req *stakingRequestedEvent) isWatched() bool {
-	return req.watchTxData != nil
-}
-
-func newOwnedStakingRequest(
-	stakerAddress btcutil.Address,
-	stakingTx *wire.MsgTx,
-	stakingOutputIdx uint32,
-	stakingOutputPkScript []byte,
-	stakingTime uint16,
-	stakingValue btcutil.Amount,
-	fpBtcPks []*btcec.PublicKey,
-	confirmationTimeBlocks uint32,
-	pop *cl.BabylonPop,
-	usePreApprovalFlow bool,
-) *stakingRequestedEvent {
-	return &stakingRequestedEvent{
-		stakerAddress:           stakerAddress,
-		stakingTxHash:           stakingTx.TxHash(),
-		stakingTx:               stakingTx,
-		stakingOutputIdx:        stakingOutputIdx,
-		stakingOutputPkScript:   stakingOutputPkScript,
-		stakingTime:             stakingTime,
-		stakingValue:            stakingValue,
-		fpBtcPks:                fpBtcPks,
-		requiredDepthOnBtcChain: confirmationTimeBlocks,
-		pop:                     pop,
-		watchTxData:             nil,
-		usePreApprovalFlow:      usePreApprovalFlow,
-		errChan:                 make(chan error, 1),
-		successChan:             make(chan *chainhash.Hash, 1),
-	}
-}
-
-type watchTxData struct {
-	slashingTx        *wire.MsgTx
-	slashingTxSig     *schnorr.Signature
-	stakerBabylonAddr sdk.AccAddress
-	stakerBtcPk       *btcec.PublicKey
-	// unbonding related data
-	unbondingTx         *wire.MsgTx
-	slashUnbondingTx    *wire.MsgTx
-	slashUnbondingTxSig *schnorr.Signature
-	unbondingTime       uint16
-}
-
-func newWatchedStakingRequest(
-	stakerAddress btcutil.Address,
-	stakingTx *wire.MsgTx,
-	stakingOutputIdx uint32,
-	stakingOutputPkScript []byte,
-	stakingTime uint16,
-	stakingValue btcutil.Amount,
-	fpBtcPks []*btcec.PublicKey,
-	confirmationTimeBlocks uint32,
-	pop *cl.BabylonPop,
-	slashingTx *wire.MsgTx,
-	slashingTxSignature *schnorr.Signature,
-	stakerBabylonAddr sdk.AccAddress,
-	stakerBtcPk *btcec.PublicKey,
-	unbondingTx *wire.MsgTx,
-	slashUnbondingTx *wire.MsgTx,
-	slashUnbondingTxSig *schnorr.Signature,
-	unbondingTime uint16,
-) *stakingRequestedEvent {
-	return &stakingRequestedEvent{
-		stakerAddress:           stakerAddress,
-		stakingTxHash:           stakingTx.TxHash(),
-		stakingTx:               stakingTx,
-		stakingOutputIdx:        stakingOutputIdx,
-		stakingOutputPkScript:   stakingOutputPkScript,
-		stakingTime:             stakingTime,
-		stakingValue:            stakingValue,
-		fpBtcPks:                fpBtcPks,
-		requiredDepthOnBtcChain: confirmationTimeBlocks,
-		pop:                     pop,
-		watchTxData: &watchTxData{
-			slashingTx:          slashingTx,
-			slashingTxSig:       slashingTxSignature,
-			stakerBabylonAddr:   stakerBabylonAddr,
-			stakerBtcPk:         stakerBtcPk,
-			unbondingTx:         unbondingTx,
-			slashUnbondingTx:    slashUnbondingTx,
-			slashUnbondingTxSig: slashUnbondingTxSig,
-			unbondingTime:       unbondingTime,
-		},
-		errChan:     make(chan error, 1),
-		successChan: make(chan *chainhash.Hash, 1),
-	}
-}
-
-func (req *stakingRequestedEvent) EventId() chainhash.Hash {
-	return req.stakingTxHash
-}
-
-func (req *stakingRequestedEvent) EventDesc() string {
-	return "STAKING_REQUESTED"
-}
 
 type stakingTxBtcConfirmedEvent struct {
 	stakingTxHash chainhash.Hash
@@ -246,20 +119,6 @@ func (app *StakerApp) logStakingEventProcessed(event StakingEvent) {
 		"eventId": event.EventId(),
 		"event":   event.EventDesc(),
 	}).Debug("Processed staking event")
-}
-
-type sendStakingTxToBTCRequestedEvent struct {
-	stakingTxHash           chainhash.Hash
-	requiredDepthOnBtcChain uint32
-	responseExpected        *responseExpectedChan
-}
-
-func (event *sendStakingTxToBTCRequestedEvent) EventId() chainhash.Hash {
-	return event.stakingTxHash
-}
-
-func (event *sendStakingTxToBTCRequestedEvent) EventDesc() string {
-	return "SEND_STAKING_TX_TO_BTC_REQUESTED"
 }
 
 type delegationActiveOnBabylonEvent struct {